Add function to parse InstanceIDs.
This function currently supports InstanceIDs in the format <Org>:<LocalID>, such as "Xen:Domain-0" for VSSD.

The functionality is based on the parsing from VSSDComponent, so the code for parsing the InstanceID is replaced by the function.

Updated function so that it can return both the InstanceID prefix (org) and the localID.  The call has the option of supplying NULL for either argument.

Also added a wrapper function that takes a CMPIObjectPath ref, grabs the instanceID, and then calls the parse_id() function.

Fixed memory leak.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 libxkutil/misc_util.c    |   49 +++++++++++++++++++++++++++++++++++++++++++++++
 libxkutil/misc_util.h    |    3 ++
 src/Virt_VSSDComponent.c |   22 +++------------------
 3 files changed, 56 insertions(+), 18 deletions(-)