Changes to misc_utils.c for connect_by_classname()
Includes Heidi's change to not return an error on failed connect.  I chose
to leave the failure in place if a proper URI could not be determined because
this really means we were called for a prefix we don't support, which /is/ a
problem.  If we need to iterate over subclasses later down the road for
differing CIMOM behavior, we should do it before we get here.

Signed-off-by: Dan Smith <danms@us.ibm.com>