Transition back to connect_by_classname()
This patch contains the previously-discussed changes to ComputerSystem and
SystemDevice (including Heidi's change to remove provider_is_responsible()).
If this looks okay to people, then I'll start in on provider-by-provider
conversions.

Previously, we used the incoming reference for all operations to determine
the URI to pass to libvirt during a connect operation.  For some reason,
we transitioned away from that to a detection-based mechanism, which would
not support multiple technologies to be managed from the provider
simultaneously.

I think that we have enough other infrastructure in place now that this is
no longer a problem.  As a test, I offer this patch with a few key instances
of lv_connect() replaced with connect_by_classname() for discussion.

I have tested that enumeration of Xen_ComputerSystem, KVM_ComputerSystem,
and CIM_ComputerSystem are exclusive, and do not return any duplicate
results.

Further, I have confirmed that resolving Xen_SystemDevice against a
KVM_ComputerSystem instance does not work, and that no duplicate results
are returned in the case of CIM_SystemDevice.

Signed-off-by: Dan Smith <danms@us.ibm.com>