Adding the RASD related part.
This is based on the Resource Allocation Profile.  My understanding is that we are implementing a setup similar to Figure 3 on page 31.  In this case, we'll use the same RASD instance for State as well as Defined.

Updated to due to change in get_rasd_instance_from_ref().  Also updated structure and handler function names to avoid confusion.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 Makefile.am               |    4 +-
 Virt_ElementSettingData.c |   70 ++++++++++++++++++++++++++++++++++++++++++----
 2 files changed, 66 insertions(+), 8 deletions(-)