Fix an issue where EnabledLogicalElementCapabilities ein/ei returns an error.
Enumerating this class currently returns: "Missing key: Name".  The problem is that get_ele_cap() expects the ref to have a Name property, which isn't the case on an ein/ei call.  This fix makes the get_ele_cap() call more generic by accepting a char * to use in building the InstanceID.

Also, the return_ele_cap() function wasn't properly returning an instance for each domain on the system.  This fix grabs the domain list and returns an instance accordingly.

NOTE:  I don't like how I handled the virDomainFree() - not sure how else to do this except for changing goto end to goto out and then looping through the list to free.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 Virt_EnabledLogicalElementCapabilities.c |   64 +++++++++++++++++++++----------
 Virt_EnabledLogicalElementCapabilities.h |    1
 2 files changed, 46 insertions(+), 19 deletions(-)