Updated to support changes to get_ele_cap().
get_ele_cap() now takes a char * argument to use for building the InstanceID.  The function needs to get the Name from the ref and pass that to get_ele_cap().  This bit of code used to be in get_ele_cap() itself, but get_ele_cap() needs to be more generic.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 Virt_ElementCapabilities.c |   16 ++++++++++++++--
 1 file changed, 14 insertions(+), 2 deletions(-)