Fix issue with VSMC enum.
This fix makes get_vsm_cap() more generic - it accepts a string to use for the InstanceID. The return_vsm_cap() handles enum'ing the VSMC instances, so this has been updated to get the hostname, which will be used in the InstanceID string.

Updated:  Instead of using gethostname() call, get the instance of the host system and use it's value for the Name attribute.  This update will allow for potiential host instances that don't use the hostname as the value for the Name attribute.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 Virt_VirtualSystemManagementCapabilities.c |   22 ++++++++++++----------
 Virt_VirtualSystemManagementCapabilities.h |    1 +
 2 files changed, 13 insertions(+), 10 deletions(-)