SD doesn't return instances in all cases.
Virt_SystemDevice doesn't return instances when CIM_ManagedElement is passed as a result class argument.  This only applies to the VM to device case (see failint query below).

The problem is that we use the result class to determine which device type to return.  If no result class is specified, we return all associated devices.  However, if CIM_ManagedElement is specified, we are unable to match the device type, and we return nothing.

I think all of this can be removed because filter_results() in libcmpiutil should remove the need for this kind of checking.

Failing query:
wbemcli ain -ac Xen_SystemDevice -arc CIM_ManagedElement 'http://root:elm3b41@localhost/root/virt:Xen_ComputerSystem.CreationClassName="Xen_ComputerSystem",Name="Domain-0"'

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 Virt_SystemDevice.c |   20 ++++----------------
 1 file changed, 4 insertions(+), 16 deletions(-)