Fix VSMS.RemoveResources() to take references instead of EOs
This just adds a filter function that converts the references into
instances.  It could be more efficient, but I think that allowing all
of the resource operations to remain largely identical (as they did before)
is worth the small amount of overhead here.

Also added REF2STR() macro to misc_util.h.  I do this a lot, mostly
while testing, but it seems valid to have such a macro so we can start
printing out more useful error messages with ease.

Signed-off-by: Dan Smith <danms@us.ibm.com>