Make NetworkPool use libvirt's virtual network API
This means that you no longer see a NetworkPool for each bridge on the
system, but rather one per configured virtual network in libvirt.  The
id is NetworkPool/NetName, and the Caption property of the instance
includes the bridge name for reference.

This change removes the need for the contents of hostres.{c,h}, so it is
removed here.  That makes the net change -85 lines of pure goodness.

Changes since last version:
 - Fixed netpool_member_of() to do extra level of bridge->network mapping

Signed-off-by: Dan Smith <danms@us.ibm.com>