Add preliminary XML-parsing support for Xen FV domains
This adds a type switch for XenPV, XenFV, KVM to the struct domain.
It also adds parsing of 'graphics' and 'emulator' devices.  Finally,
amidst the compulsory changes to xmlgen.c are modifications to
generate a proper <os> block for XenFV domains.  More work will need
to be done here, but this is an example of how the domain->type switch
will be used to select different generation routines.

Changes:
 - Make VIRT_DEV_UNKNOWN start after CIM types, and with a constant
   so that automatic numbering starts again after the CIM range
 - Set the device types in the parse_*_device() functions for emu and graphics
 - Removed stale debug statement
 - Cleaned up free() of type string in cleanup_dominfo()

Signed-off-by: Dan Smith <danms@us.ibm.com>