Add XML parsing tester for use when working on Xen PV and KVM support
This simply takes a domain argument, gets the XML from libvirt, parses
it in the same way the providers will, and prints the domain information
in a structured way.  This can be used to test the previous patch, by
validating the FV-specific data items, such as emulator and the HVM loader
path.

Signed-off-by: Dan Smith <danms@us.ibm.com>