.#2 Fix get_graphics_device() - returns error if graphics tag is missing.
Update:  Instead of returning a constant 1 from get_graphics_device() and get_emu_device(), keep the return value as is.  Don't capture the return value of these functions in get_dominfo().

The graphics tag isn't mandatory, so we shouldn't capture an error in get_dominfo(). The same problem exists with get_emu_device().

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 device_parsing.c |    4 ++--
 1 file changed, 2 insertions(+), 2 deletions(-)