Allow device_parsing modules to return multiple devices per cycle
Do this by growing the device list during the cycle instead of
trying to predict it ahead of time based on the numbers of nodes
in the XML tree.  This could be a little nicer with a more integrated
list structure or a linked-list of some kind, but this is the
smallest change to accomplish the goal, which I think is a good idea
for now.

The need for this is apparent in the next patch for vcpu parsing.

Signed-off-by: Dan Smith <danms@us.ibm.com>