Make device_parsing get XML description for a domain once for all phases
Also, make memory devices parse the XML instead of ops on a domain.  This
gives us the ability to create a dominfo structure from XML-only.

It would have been nice to put this into two separate patches, but I would have had to fake some memory stuff just to make the reorg work and then remove it
in the next patch, so hopefully what I did is palatable to all :)

Signed-off-by: Dan Smith <danms@us.ibm.com>