Update VSSD schemas for FullVirt
This adds a flag to the Xen VSSD to enable full-virt or not, as well as the
required boot device information.  In the implementation, a "false" or
missing value for the flag will indicate paravirt-ness.

For KVM, I added the boot device as well, since that's something it will
always need.  For both, a missing boot device will imply "hd".

Changes:
- Fixed capitalization of IsFullVirt

Signed-off-by: Dan Smith <danms@us.ibm.com>