Protect CSI from transient domains
Testing has revealed a couple of ways to segfault the CSI provider.  One is a domain that is undefined immediately after being defined, the other is a domain that is listed as defined by libvirt but has not actually been fully defined by xend.  In either case, the problem boils down to CSI trying to get an XML description of a domain that isn't there.  This eventually results in a segfault.

Signed-off-by: Jay Gagnon <grendel@linux.vnet.ibm.com>

 Virt_ComputerSystemIndication.c |   30 ++++++++++++++++++++++--------
 1 file changed, 22 insertions(+), 8 deletions(-)