Fix bizarre crash in VSMigrationService
Apparently I was the only one seeing this, but I hit a very strange bug in Virt_VSMigrationService.  It would just totally implode whenever I tried to migrate a system.  It was eventually tracked down to the transport member of the migration_job struct; for some reason if that member (which we don't even use anymore) is removed everything works perfectly.  This might bear more investigation, but for the time being the fix should at least be put in.

Signed-off-by: Jay Gagnon <grendel@linux.vnet.ibm.com>

 Virt_VSMigrationService.c |    1 -
 1 file changed, 1 deletion(-)