Make VSMS not parse the InstanceID of a RASD, but rather use the proper fields
For Network, grab the Address for the MAC and then also try to extract
the PoolID for the network name.  I left it split between KVM and Xen,
because I think as we expand the supported network types, we'll still need
to differentiate in the code.

This means that we no longer examine the InstanceID of the incoming RASDs
for a DefineSystem operation (which is correct).  The test suite will need
to change to use these values correctly instead of generating the InstanceID
and should also have a test that verifies that the InstanceID is not
honored.  I think that the InstanceID should be ignored if specified,
although discussion of this point is certainly welcomed.

Changes:
 - Use exposed defaults for network pool (but not the MAC)
 - Set the namespace on the instance of the RASD we got from the parser,
   so that it's a usable reference for namespace, etc.

Signed-off-by: Dan Smith <danms@us.ibm.com>