Add revision information to the VSMS
This is done at build time, if possible.  If not, the changeset/revision
at the last automake run will be used.  This should make it work for the
tarball/release case as well as the RPM build.

The way this works, if you build, then apply a patch or pull a new changeset,
the exposed revision/version information will be stale unless you do a clean.
Since this is really just for the purposes of test-tracking, I think this
is a reasonable restriction that can be minded by the people doing testing.

The changeset hash and revision number are exposed in aptly-named properties
of the VSMS for each platform.  If anyone has a better place for these to
show up, or a better format, please bring it up.

Changes:
 - Update the VSMS schema for the new values

Signed-off-by: Dan Smith <danms@us.ibm.com>