(#3) Fix device_parsing of processors.
Kaitlin discovered that my changes to device_parsing.c make it so we only get one Processor as well, not just one Proc RASD.  This moves the flattening step up a little higher so that enumerating on Processor works correctly.  It's not the prettiest approach to things, but I think given Processor's unique Device-RASD relationship it's a good compromise.  I would of course be open to suggestions for making it a bit more elegant.

Changes from 1 to 2:
Fix xml_parse_test.c's print_dev_vcpu, and by "fix" I mean "fix, because once again I made a change that breaks it."

Changes from 2 to 3:
VirtualQuantity is supposed to be 64-bit, not 32.

Signed-off-by: Jay Gagnon <grendel@linux.vnet.ibm.com>

 libxkutil/device_parsing.c |   48 ++++++++++++++-------------------------------
 libxkutil/device_parsing.h |    2 -
 libxkutil/xml_parse_test.c |    4 ++-
 src/Virt_RASD.c            |   20 +++++++++++++++++-
 4 files changed, 38 insertions(+), 36 deletions(-)