[RFC] Reorder providers in Mafefile.am.
Currently, the Virt_VSMigrationCapabilities library is not being installed properly.

/usr/bin/ld: cannot find -lVirt_VSMigrationCapabilities^M
collect2: ld returned 1 exit status^M
libtool: install: error: relink `libVirt_SettingsDefineCapabilities.la' with the above command before installing it

During the install, libtool has to relink libraries once they are in the install location.  Here, libtool attempts relink the library before the dependant libraries are installed.

This change reorders things so that class libraries are listed first in the list and association libraries are listed at the end.  Also, the class libraries are reordered so that those libraries with dependancies are listed after the libraries they are dependant on.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>

 Makefile.am |   38 ++++++++++++++++++++------------------
 1 file changed, 20 insertions(+), 18 deletions(-)