Add the HostSystem-to-AC link with EC
This happens as part of the existing HostSystem->*Capabilities code,
but we also insert the Characteristics array into the Reference instance.
As far as I can tell, this is okay to be 2=Default for all those references.
Also, only do this for the MemoryAC at the moment, since only that profile
specifies this behavior.

The reverse direction replaces the unsupported AC-to-pool code with
bits to return the HostSystem.

Changes:
 - Make alloc_to_pool() named something more appropriate

Signed-off-by: Dan Smith <danms@us.ibm.com>