(#4) ProcessorRASD, the class that won't go away
Around and round we go on the merry-go-round of indecision, hopefully for our last ride.  The definitely probably absolutely tentative final plan is that the default representation of processors as virt_device structs will be one per domain, with a quantity.  Virt_Devices will just need to be told how to handle that, and all the RASD stuff will be much happier for it.  And then we will never speak of this again.

So today being my last official day on the team, I kind of have to get this out in a not perfect state.  From what I can see it does what it is supposed to do, but it is entirely possible (likely?) that there are a few things in here that aren't really great things to do.  I wish I could polish it more but I think I've at least got it to the point where any work left to be done is generic "not a great idea in C" problems, as opposed to the much more annoying "what exactly are we supposed to do in this case" problems.

Changes from 1 to 2:
Close brace on if block.  That one is really embarrassing; I have no idea what happened there.  I mean that crap won't even compile.
Removed unnecessary proc_found boolean.  That was me being lazy after changing my mind on something.
Cleaned up a couple leftover CU_DEBUG calls.
Addition of vcpu_inst function, which is similar to vcpu_instances but always only works on one instance and takes a dev_id.  This is to prevent all instances from being given a dev_id of zero.  It is unfortunately similar to vcpu_instances in other regards, but I'm not sure if it's worth the overhead of trying to combine the similar parts.
Fixed quantity v number problem in xml_parse_test that I appear to have been failing to address for about six years.

Changes from 2 to 3:
Actually include changes needed for associations.
Make vcpu_instances use vcpu_inst.

Changes from 3 to 4:
Rename list_devs to something more meaningful
Fix misplaced free.

Signed-off-by: Jay Gagnon <grendel@linux.vnet.ibm.com>
Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>
Signed-off-by: Dan Smith <danms@us.ibm.com>