Allow KVM and Xen to parse both network and bridge interfaces.
Since KVM and Xen guests are both created with network interfaces, there no longer needs to be a seperate function for KVM and Xen.

This change allow fixes an error seen when attempting to modify a KVM guest that was created with a non-network inteface (created using a method other than the providers).  Even though we don't create guests with bridge type interfaces, we should be able to parse their xmls.

This bit of code will probably need to be updated to support other interface types.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>