Fix race between EnableIndications and ActivateFilter's thread
Accomplish this by decoupling the enabled-ness of the whole provider from
the presence of activated filters against the provider.  Keep a count of
the number of subscribed filters against us (per platform) and keep the
thread running until that reaches zero.  Use the enabled-ness of the
provider to gate actual delivery of any indications (as intended).

The original race was because ActivateFilter could be run before
EnableIndications (which seems broken to me -- thanks Pegasus) and thus
the thread was started before the global flag was set true.  Thus, the
thread fell right through the while loop and stopped running.

Signed-off-by: Dan Smith <danms@us.ibm.com>