Add ReferenceConfiguration support to VSMS
This accomplishes that goal by doing a full get_dominfo() instead of
a blank dominfo alloc in the event that the client passed in a
ReferenceConfiguration object path.

Also fix a few unchecked calloc() calls while touching them.  Also
make sure we're only ever editing the first memory or processor device
in the list, during the domain building process.

Signed-off-by: Dan Smith <danms@us.ibm.com>