(#4) For pause, reboot, and enable add support for guests in the NOSTATE state.
It's possible that a XenFV guest might return no state instead of blocked or running.  This can occur when the guest is active, but isn't running of the CPU or blocked waiting for it.

This should be a safe approach - we return an error if the guest is in some other state or if libvirt is unable to successfully change the state of the guest.

Updates from 3 to 4:
  -Use single function to convert Xen NOSTATE to RUNNING.

Updates from 2 to 3:
  -Add logic to check connection type so NOSTATE only applied to Xen guests.

Updates from 1 to 2:
  -For the reboot and reset operations, add support for paused guests (per the VSP).
  -If the guest returns VIR_DOMAIN_NOSTATE, the provider should return the value CIM_STATE_ENABLED, not CIM_STATE_UNKNOWN.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>