Override CreateSnapshot extrinsic to include vendor-defined values
Some tools will parse the class info and format gui based on qualifiers
it finds.  In the case of CreateSnapshot, such tools will not display
the vendor additions unless the method is overridden to describe the
additions.

This patch introduces a new class Virt_VirtualSystemSnapshotService
where CreateSnapshot is overridden.  The various virtualizer flavors
then inherit from that.

     Signed-off-by: Jim Fehlig <jfehlig@novell.com>