(#2) Override devices in reference configuration, where the ID is the same
Currently, we depend on libvirt to return an error if the client's
combination of ReferenceConfig and ResourceSettings results in a duplicate
device.  That error is non-obvious in origin.  Since two devices that would
result in the same InstanceID cannot be properly represented by the providers
anyway, this patch adds some logic to properly override such a conflicting
device.

To do this, I had to make sure we set the ID in rasd_to_vdev, which wasn't
previously required.  The actual insertion of the device into the appropriate
type list is done by a helper function which checks for (and overrides if
appropriate) a duplicate before tacking the new instance on the end of the
list.

Changes:
We were already making the required space for the additional devices,
but I didn't correctly specify the new size then calling add_device_nodup()
function, resulting in a false failure.

Side note: the classify_resources() function is really getting out of hand
and could use some refactoring.

Signed-off-by: Dan Smith <danms@us.ibm.com>