(#2) Add ConcreteComponent with NetworkPool support
This adds the ConcreteComponent association, and hooks it up to NetworkPool
for now.  If you have sblim-cmpi-network on your system, this references
the Linux_EthernetPort instance exposed by SBLIM for the bridge device
associated with the network pool.

Note that port_to_netpool() is untested and won't get called by the CIMOM
because we're not registering this association in root/cimv2.  I don't
think it makes much sense for us to do that to support this backwards,
so I'm content to leave this as kinda broken until someone comes up with a
better idea, or complains.

Changes:
 - Match the hypervisor prefix to avoid duplicate results
 - Rename GroupComponent and PartComponent to be more appropriate

Signed-off-by: Dan Smith <danms@us.ibm.com>