Remove the default implied VirtualDevice on DiskRASD
I'm not sure what I was thinking when I allowed the client to not specify
a VirtualDevice for a given disk, but this patch removes that and turns
it into an error.

The Address field needs to have similar semantics, but will be a little
more complicated for CDROM and thus will be handled in the upcoming set.

Signed-off-by: Dan Smith <danms@us.ibm.com>