(#2) Remove CMSetObjectPath() from rasd_to_vdev()
This is only needed because default_device_pool() needs an ObjectPath with a proper namespace set. Since this is only needed in the network case, remove the calls to CMSetObjectPath() and CMSetNameSpace() from rasd_to_vdev().

Instead, pass the namespace to _default_network(), which sets the namespace on the ObjectPath directly.  This removes any need of setting the ObjectPath of the instance.

Also, remove kvm_net_rasd_to_vdev() and xen_net_rasd_to_vdev() - these calls are identical.  Move this functionality into net_rasd_to_vdev() and clean up net_rasd_to_vdev().  This means that only net_rasd_to_vdev() needs to be updated with an additional parameter.

Updates:
  -Added ns paramerter to _container_rasd_to_vdev() for consistency with _sysvirt_rasd_to_vdev()

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>