Remove sdc_rasd_prop structs from SDC
Instead, a pointer to CMPIInstance can be passed to the <>_template() functions, and the functional can be responsible for setting the properties directly.  Also, the functions will now return a CMPIStatus value instead of a sdc_rasd_prop struct.

This is more useful than structs because some of the property values may differ between the virtualization types.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>