Add KVMRedirectionSAP provider.
The profile calls for a distinction between an active session (where the redirection is currently being used) and an available session (a redirection is in place, but not currently in use).

There's not an easy way to detect this, so the session will be set to active all the time.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>