Add an EmulatedType field to DiskRASD to select CDROM or Disk
This seems like a pretty reasonable way to do this, but comments are welcome.
I had initially planned to have a specific RASD type to represent a CDROM,
but I don't think that makes much sense, and is significantly more complex.
Adding this gives us a way to set and query the CDROM-ness of a disk, and
with the following patch, avoids dropping this qualifier from existing
configurations.

Signed-off-by: Dan Smith <danms@us.ibm.com>