Add vir_set_status() to misc_util.h
This function allows you to set the message of a CMPIStatus object to something
like what perror() would do.  You give it an error message and (optionally)
a virConnectPtr object and it formats it as "YourError: VirtError".  It
attempts to extract the last error from the connection pointer if supplied,
otherwise it takes the last error in the global library pointer.

I tried to make this function as foolproof as possible, providing no real
error paths.  It attempts to get as much info as it can into the status
object, despite any failures along the way.  Since this is usually to be used
for error handling scenarios, this seems like the only sane approach.

Signed-off-by: Dan Smith <danms@us.ibm.com>