Make DefineSystem() and RequestStateChange() provide rich errors
This patch adds use of the new virt_set_status() function to a couple of
providers where it's likely to be the most effective.  A DefineSystem()
call can fail for a variety of reasons, as can an attempt to change the
state of a guest.  By pulling in the libvirt error message, we're much more
likely to be able to debug the issue from just the error message the CIM
client gets.

Signed-off-by: Dan Smith <danms@us.ibm.com>