Catch unimplemented sentinel of GetMaxVcpus and be less stupid
Currently, we stuff the result of GetMaxVcpus into a u64 directly.  Since
that function can return -1 for unimplemented (as it does for LXC), we end
up with a ridiculously large number of vcpus as the maximum.

Assuming nothing can run with less than 1 cpu, I think it's a safe default
for this case.

Signed-off-by: Dan Smith <danms@us.ibm.com>