Replace CBGetInstance() with CBEnumInstances() in Virt_HostSystem
This is a work around to deal with the way the SBLIM provider determines the
FQDN of a host system.  The value the SBLIM provider returns can depend on
how the /etc/hosts file is formatted, as well as other factors (running
dnsdomainname caused the provider to print two different values on my system).

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>