If the resource type of a devices is set to UNKNOWN, do not generate XML for it
xmlgen was recently refactored, so this adds back in behavior that was missed
during the refactor.  If the resource type of a device is CIM_RES_TYPE_UNKNOWN,
then the xml generation for that device should be skipped.

The RemoveResources() call sets a device's type to UNKNOWN to indicate that
the device should be marked for removal.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>