Add ForwardDevice attribute for network pools
libvirt allows you to restrict the firewall rules of a pool so that traffic
is forwarded through a specific device.

Also, fix a bug where the value for ForwardMode specified by the user wasn't
being read in properly.  Also change FowardMode to a value map attribute
instead of a string.

Fix a bug where the value for IPRangeStart is read in twice - IPRangeEnd wasn't
being read in at all.

Fix a bug in xmlgen code for the network pool forwarding device.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>