Remove migration related super classes
These classes are now a part of the 2.21 schema. Also remove the TransportType
attribute - this is part of the VirtualSystemMigrationSettingData class.

The valuemap has changed, so the UNIX transport type needs to have a value of
32768 instead of 6.

Add new parameters to migration methods - make these optional since we don't
currently use them.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>