Add support for logical pools to SDC
There isn't much that needs to be done for logical pools.  The user will
need to specify the Path as something like:  "/dev/VolGroup01"

When calling CreateChildResourcePool(), the user will also need to specify
the ElementName as an existing LVM volume group, like: "VolGroup01". This is
different from the other pools (which any name for ElementName can be used).

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>