Remove type_name param from _do_indication in CSIndication
No reason to pass the indication name if we're already passing the indication
type.

Also, remove the call to set the AffectedSystem property.  This property doesn't
exist in any of the Indications classes in the 2.21 schema - also doesn't
exist in our own local schema.  Not sure why this value was being set, doesn't
appear to be a property in the 2.16 schema either.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>