Add raise_indication() to CSIndication
Currently, the event loop in ComputerSystemIndication gets the XML of the guest
when the guest has been modified.  However, by the time the event loop is
called, the the guest has already been modified.  This is because we only
generate the indication if the modification is successful.

Therefore, we need a function that allows us to raise an indication (instead of
trigger in this case).  This allows us to set the PreviousInstance attribute of
the indication instance we pass to the raise call.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>