Fix possible seg fault in resource_add() and resource_mod()
If a user attempts to add a virtual NIC with a MAC that is already in use, the
provider doesn't current check the return from rasd_to_vdev().  This means
the net device struct isn't properly filled out and will result in a seg fault.

To fix this, be sure to check the return of rast_to_vdev() and return an error
accordingly.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>