(#2) Add LibvirtVersion to VSMS
I've had users ask for libvirt version available through libvirt-cim.
This patch add LibvirtVersion property to VSMS, which is a proxy for
libvirt and seems the most appropriate place to add the property.

I consider VSMSC, but that class describes what the associated service
is capable of as opposed to version information about the service.

#2:
If unable to determine libvirt version, set property to
"Unknown libvirt version" instead of "Unknown libvirt".

     Signed-off-by: Jim Fehlig <jfehlig novell com>