Fix logic for checking UUID conflicts in ModifySystemSettings()
Instead of checking to see if the UUID is in use, we need to make sure the
provider is using the existing UUID.  If the user specifies a UUID that is
different, then an error is returned.  If no UUID (or if an empty string is
specified), the provider will override that value with the original UUID.

This fixes a bug where the user specifies a empty string, which we were passing
to libvirt.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>