(#2) This patch adds support for IPv6 in KVMRedirectionSAP provider.
The code now loops through currently active IPv4 and IPv6 connections.
It tries to read both /proc/net/tcp and /proc/net/tcp6 files. If one
of the two files is successfully read, it does not return an error. For
the function to return an error, both IPv4 and IPv6 file reads should fail.

Signed-off-by: Sharad Mishra <snmishra@us.ibm.com>