Fix storage volume generation so that bytes are supported
If the units portion of the <capacity> tag isn't specified, then libvirt
will create the storage volume using bytes as the unit type.  libvirt-cim
wasn't supporting this behavior properl.  Instead of assuming gigabytes as the
default, libvirt will assume bytes when no unit is specified.

If a unit is specified, it needs to be one of the following supported types:
 'k', 'K', 'm', 'M', 'g', 'G', 't', 'T', 'p', 'P', 'y', or 'Y'.

Signed-off-by: Kaitlin Rupert <karupert@us.ibm.com>