(#2) Do not filter out SBLIM host.
This patch removes the code that does not return SBLIM host system.
The original code did not have this filter and was added later.
I searched the archives and did not find any reason to add this code.

Update #2 - Added an else case where if enumeration of instances is null,
            a fake host is returned.

Signed-off-by: Sharad Mishra <snmishra@us.ibm.com>