Remove test for duplicate mac address.
This test has been removed since it was interfering with modifying network resources. This is an extra piece of check that was added in libvirt-cim, libvirt already does this check. Removing it should not cause any problem.
When a user modified an existing active net RASD, libvirt-cim clones the RASD and changes the properties to the new values. But if mac was not one of the changing property, then this extra check, took it to be a duplicate mac and gave error.

Signed-off-by: Sharad Mishra <snmishra@us.ibm.com>