(#2) Set CreationClassName property in VSMS
CCName was not getting set correctly. I do not understand why there was a check to see if CCName is set, and set it only if it is set.
This patch removed the check, and sets the ccname if CMPIInstance is not null and ccname property is non-null.

update #2: Removed unused 'data' variable.

Signed-off-by: Sharad Mishra <snmishra@us.ibm.com>